+++ CHANGELOG.md
@@ -32,6 +32,8 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids wrong annotations in assignment
 - Forbids using multiline `for` and `while` statements
 - `WPS113` now can be tweaked with `I_CONTROL_CODE` setting
+- Adds `WPS000` that indicates internal errors
+- Forbids to use implicit `yield from`
 
 ### Bugfixes